Remove SKE feature flag
