+++ app/views/provider_interface/offer/conditions/_form.html.erb
@@ -40,11 +40,7 @@
           tabindex: '-1',
         },
       ) do %>
-      <% if FeatureFlag.active?(:provider_ske) %>
-        <p class="govuk-body">For example, completing their degree with a certain grade.</p>
-      <% else %>
-        <p class="govuk-body">For example, studying a subject knowledge enhancement course.</p>
-      <% end %>
+      <p class="govuk-body">For example, completing their degree with a certain grade.</p>
 
         <%= render 'provider_interface/offer/conditions/further_condition', form: f, model: Struct.new(:id, :condition_id, :text).new('placeholder', nil, nil), label_text: 'placeholder', disabled: true %>
