+++ spec/system/provider_interface/make_offer_ske_language_flow_spec.rb
@@ -1,6 +1,6 @@
 require 'rails_helper'
 
-RSpec.feature 'Provider makes an offer with SKE enabled in standard courses' do
+RSpec.feature 'Provider makes an offer with SKE enabled on language flow' do
   include DfESignInHelpers
   include ProviderUserPermissionsHelper
   include OfferStepsHelper
@@ -20,17 +20,15 @@ RSpec.feature 'Provider makes an offer with SKE enabled in standard courses' do
   let(:course_option) { build(:course_option, course:) }
 
   before do
-    given_the_course_subject_requires_ske
+    given_the_course_subject_is_modern_language
   end
 
   scenario 'Making an offer for the requested course option' do
     given_i_am_a_provider_user
     and_i_am_permitted_to_make_decisions_for_my_provider
     and_i_sign_in_to_the_provider_interface
-    and_provider_ske_feature_flag_is_enabled
 
     given_the_provider_has_multiple_courses
-    and_the_other_courses_subject_does_not_require_ske
     given_the_provider_user_can_offer_multiple_provider_courses
 
     when_i_visit_the_provider_interface
@@ -39,33 +37,47 @@ RSpec.feature 'Provider makes an offer with SKE enabled in standard courses' do
     then_i_see_the_decision_page
 
     when_i_choose_to_make_an_offer
+    then_the_ske_language_flow_is_loaded
 
-    then_the_ske_standard_flow_is_loaded
-    when_i_dont_select_any_ske_answer
-    then_i_should_see_a_error_message_to_select_if_ske_required
-
-    when_i_select_no_ske_required
+    when_i_select_no_language
     and_i_click_continue
     then_the_conditions_page_is_loaded
 
     and_i_click_back
 
-    when_i_select_ske_is_required
+    then_the_ske_language_flow_is_loaded
+    when_i_dont_select_any_ske_answer
+    then_i_should_see_a_error_message_to_select_language
+
+    and_i_select_language_and_the_no_option
+    and_i_click_continue
+    then_i_should_see_a_error_message_to_select_one_or_the_other_language
+
+    when_i_select_three_languages
+    and_i_click_continue
+    then_i_should_see_an_error_message_to_select_no_more_than_2_languages
+
+    when_i_select_two_languages
     and_i_click_continue
     then_the_ske_reason_page_is_loaded
 
     when_i_dont_give_a_ske_reason
-    then_i_should_see_a_error_message_to_give_a_reason_for_ske
+    then_i_should_see_a_error_message_to_give_a_reason_for_ske_for_all_languages
 
-    when_i_add_a_ske_reason
+    when_i_add_a_ske_reason_for_all_languages
     and_i_click_continue
     then_the_ske_length_page_is_loaded
 
     when_i_dont_answer_ske_length
-    then_i_should_see_a_error_message_to_give_a_ske_course_length
+    then_i_should_see_a_error_message_to_give_a_ske_course_length_for_all_languages
+
+    when_i_select_both_ske_courses_over_8_weeks
+    and_i_click_continue
+    then_i_should_see_a_error_message_to_select_at_least_one_8_week_course
 
     when_i_answer_the_ske_length
     and_i_click_continue
+
     then_the_conditions_page_is_loaded
     and_the_default_conditions_are_checked
 
@@ -74,6 +86,7 @@ RSpec.feature 'Provider makes an offer with SKE enabled in standard courses' do
     and_i_click_continue
     then_the_review_page_is_loaded
     and_i_can_confirm_my_answers
+    and_the_ske_conditions_should_be_displayed
 
     when_i_click_change_course
     then_i_am_taken_to_the_change_course_page
@@ -113,31 +126,59 @@ RSpec.feature 'Provider makes an offer with SKE enabled in standard courses' do
     then_i_see_that_the_offer_was_successfuly_made
   end
 
-  def given_the_course_subject_requires_ske
+  def given_the_course_subject_is_modern_language
     application_choice.course_option.course.subjects.delete_all
     application_choice.course_option.course.subjects << build(
-      :subject, code: 'C1', name: 'Biology'
+      :subject, code: '15', name: 'Portuguese'
     )
   end
 
-  def and_the_other_courses_subject_does_not_require_ske
-    @provider_available_course.subjects.delete_all
-    @provider_available_course.subjects << build(
-      :subject, code: 'W1', name: 'Art and design'
-    )
-  end
-
-  def then_the_ske_standard_flow_is_loaded
+  def then_the_ske_language_flow_is_loaded
     expect(page).to have_current_path("/provider/applications/#{application_choice.id}/offer/ske-requirements/new", ignore_query: true)
   end
 
   def when_i_dont_select_any_ske_answer
+    uncheck 'No, a SKE course is not required'
     click_on 'Continue'
   end
 
-  def then_i_should_see_a_error_message_to_select_if_ske_required
+  def then_i_should_see_a_error_message_to_select_language
     expect(page).to have_content('There is a problem')
     expect(page).to have_content('Select whether you require the candidate to do a course')
+  end
+
+  def and_i_select_language_and_the_no_option
+    check 'French'
+    check 'No, a SKE course is not required'
+  end
+
+  def then_i_should_see_a_error_message_to_select_one_or_the_other_language
+    expect(page).to have_content('Select a language, or select ‘No, a SKE course is not required’')
+  end
+
+  def when_i_select_no_language
+    check 'No, a SKE course is not required'
+  end
+
+  def when_i_select_three_languages
+    check 'French'
+    check 'Spanish'
+    check 'German'
+  end
+
+  def then_i_should_see_an_error_message_to_select_no_more_than_2_languages
+    expect(page).to have_content('Select no more than 2 languages')
+  end
+
+  def when_i_select_two_languages
+    uncheck 'German'
+    uncheck 'No, a SKE course is not required'
+    check 'French'
+    check 'Spanish'
+  end
+
+  def then_the_ske_reason_page_is_loaded
+    expect(page).to have_current_path("/provider/applications/#{application_choice.id}/offer/ske-reason/new", ignore_query: true)
   end
 
   def when_i_select_no_ske_required
@@ -160,13 +201,19 @@ RSpec.feature 'Provider makes an offer with SKE enabled in standard courses' do
     click_on 'Continue'
   end
 
+  def then_i_should_see_a_error_message_to_give_a_reason_for_ske_for_all_languages
+    expect(page).to have_content('There is a problem')
+    expect(page).to have_content('Select why the candidate needs to take a course')
+  end
+
   def then_i_should_see_a_error_message_to_give_a_reason_for_ske
     expect(page).to have_content('There is a problem')
     expect(page).to have_content('Select why the candidate needs to take a course')
   end
 
-  def when_i_add_a_ske_reason
-    choose t('provider_interface.offer.ske_reasons.different_degree', degree_subject: application_choice.current_course.subjects.first.name)
+  def when_i_add_a_ske_reason_for_all_languages
+    choose t('provider_interface.offer.ske_reasons.different_degree', degree_subject: 'French')
+    choose t('provider_interface.offer.ske_reasons.different_degree', degree_subject: 'Spanish')
   end
 
   def then_the_ske_length_page_is_loaded
@@ -177,13 +224,42 @@ RSpec.feature 'Provider makes an offer with SKE enabled in standard courses' do
     click_on 'Continue'
   end
 
+  def when_i_select_both_ske_courses_over_8_weeks
+    form_groups.first.choose('28 weeks')
+    form_groups.last.choose('12 weeks')
+  end
+
+  def then_i_should_see_a_error_message_to_give_a_ske_course_length_for_all_languages
+    expect(page).to have_content('There is a problem')
+    expect(page).to have_content('Select how long the course must be')
+  end
+
   def then_i_should_see_a_error_message_to_give_a_ske_course_length
     expect(page).to have_content('There is a problem')
     expect(page).to have_content('Select how long the course must be')
   end
 
+  def then_i_should_see_a_error_message_to_select_at_least_one_8_week_course
+    expect(page).to have_content('There is a problem')
+    expect(page).to have_content('Select one language course that’s 8 weeks, the other course can be between 8 and 28 weeks')
+  end
+
   def when_i_answer_the_ske_length
-    choose '8 weeks'
+    form_groups.first.choose('8 weeks')
+    form_groups.last.choose('12 weeks')
+  end
+
+  def form_groups
+    page.all('.govuk-form-group')
+  end
+
+  def and_the_ske_conditions_should_be_displayed
+    %w[French Spanish].each do |language|
+      expect(page).to have_content('Subject knowledge enhancement course')
+      expect(page).to have_content("Subject\n#{language}")
+      expect(page).to have_content("Length\n12 weeks")
+      expect(page).to have_content("Reason\nTheir degree subject was not #{language}")
+    end
   end
 
   def subject_name
