+++ spec/system/provider_interface/make_offer_spec.rb
@@ -7,21 +7,21 @@ RSpec.feature 'Provider makes an offer' do
   let(:provider_user) { create(:provider_user, :with_dfe_sign_in) }
   let(:provider) { provider_user.providers.first }
   let(:ratifying_provider) { create(:provider) }
-  let(:application_form) { build(:application_form, :minimum_info) }
+  let(:application_form) { build(:application_form, :minimum_info, recruitment_cycle_year: 2023) }
   let!(:application_choice) do
     create(:application_choice, :awaiting_provider_decision,
-           application_form:,
-           course_option:)
+           application_form: application_form,
+           course_option: course_option)
   end
   let(:course) do
-    build(:course, :full_time, provider:, accredited_provider: ratifying_provider)
+    build(:course, :full_time, provider: provider, accredited_provider: ratifying_provider)
   end
-  let(:course_option) { build(:course_option, course:) }
+  let(:course_option) { build(:course_option, course: course) }
 
   scenario 'Making an offer for the requested course option' do
     given_i_am_a_provider_user
     and_i_am_permitted_to_make_decisions_for_my_provider
-    and_provider_ske_feature_flag_is_disabled
+    and_provider_structured_reference_condition_is_disabled
     and_i_sign_in_to_the_provider_interface
 
     given_the_provider_has_multiple_courses
@@ -35,6 +35,7 @@ RSpec.feature 'Provider makes an offer' do
     when_i_choose_to_make_an_offer
     then_the_conditions_page_is_loaded
     and_the_default_conditions_are_checked
+    and_the_reference_guidance_is_visible
 
     when_i_add_further_conditions
     and_i_add_and_remove_another_condition
@@ -88,8 +89,8 @@ RSpec.feature 'Provider makes an offer' do
     permit_make_decisions!
   end
 
-  def and_provider_ske_feature_flag_is_disabled
-    FeatureFlag.deactivate(:provider_ske)
+  def and_provider_structured_reference_condition_is_disabled
+    FeatureFlag.deactivate(:structured_reference_condition)
   end
 
   def and_i_sign_in_to_the_provider_interface
@@ -125,6 +126,12 @@ RSpec.feature 'Provider makes an offer' do
   def and_the_default_conditions_are_checked
     expect(find("input[value='Fitness to train to teach check']")).to be_checked
     expect(find("input[value='Disclosure and Barring Service (DBS) check']")).to be_checked
+  end
+
+  def and_the_reference_guidance_is_visible
+    expect(page).to have_content('The candidate will confirm which references they want to request when they accept your offer.')
+    expect(page).to have_content('They’ll be told they need 2 references including:')
+    expect(page).to have_content('an academic tutor if they have graduated in the past 5 years or are still studying the headteacher if they’ve been working in a school')
   end
 
   def when_i_add_further_conditions
@@ -174,8 +181,8 @@ RSpec.feature 'Provider makes an offer' do
   end
 
   def given_the_provider_has_multiple_courses
-    @provider_available_course = create(:course, :open_on_apply, study_mode: :full_time, provider:, accredited_provider: ratifying_provider)
-    create(:course, :open_on_apply, provider:)
+    @provider_available_course = create(:course, :open_on_apply, study_mode: :full_time, provider: provider, accredited_provider: ratifying_provider)
+    create(:course, :open_on_apply, provider: provider)
     course_options = [create(:course_option, :full_time, course: @provider_available_course),
                       create(:course_option, :full_time, course: @provider_available_course),
                       create(:course_option, :full_time, course: @provider_available_course)]
@@ -210,7 +217,7 @@ RSpec.feature 'Provider makes an offer' do
 
   def given_the_provider_user_can_offer_multiple_provider_courses
     @available_provider = create(:provider)
-    create(:provider_permissions, provider: @available_provider, provider_user:, make_decisions: true)
+    create(:provider_permissions, provider: @available_provider, provider_user: provider_user, make_decisions: true)
     courses = [create(:course, study_mode: :full_time_or_part_time, provider: @available_provider, accredited_provider: ratifying_provider),
                create(:course, :open_on_apply, study_mode: :full_time_or_part_time, provider: @available_provider, accredited_provider: ratifying_provider)]
     @selected_provider_available_course = courses.sample
@@ -223,14 +230,14 @@ RSpec.feature 'Provider makes an offer' do
     create(
       :provider_relationship_permissions,
       training_provider: provider,
-      ratifying_provider:,
+      ratifying_provider: ratifying_provider,
       ratifying_provider_can_make_decisions: true,
     )
 
     create(
       :provider_relationship_permissions,
       training_provider: @available_provider,
-      ratifying_provider:,
+      ratifying_provider: ratifying_provider,
       ratifying_provider_can_make_decisions: true,
     )
